# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Productos",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
        "/api/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Mostrar productos",
                "description": "Retorna la lista de productos",
                "operationId": "App\\Http\\Controllers\\ProductoController::getProductos",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductoResource"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear Producto",
                "description": "Crea un nuevo producto",
                "operationId": "CrearProducto",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductoResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/productos/{id}": {
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Modificar Producto",
                "description": "Modifica un producto existente",
                "operationId": "ModificarProducto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductoResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/productos/delete/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar producto",
                "description": "Elimina un producto",
                "operationId": "App\\Http\\Controllers\\ProductoController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductoResource"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateProductoRequest": {
                "title": "ProductoRequest",
                "description": "Producto request body data",
                "required": [
                    "codigo",
                    "nombre"
                ],
                "properties": {
                    "codigo": {
                        "title": "codigo",
                        "description": "codigo del producto",
                        "type": "string"
                    },
                    "nombre": {
                        "title": "nombre",
                        "description": "nombre del producto",
                        "type": "string"
                    },
                    "descripcion": {
                        "title": "descripcion",
                        "description": "descripcion del producto",
                        "type": "string"
                    },
                    "cantidad": {
                        "title": "cantidad",
                        "description": "cantidad del producto",
                        "type": "number"
                    }
                },
                "type": "object"
            },
            "Producto": {
                "title": "Producto",
                "description": "Producto model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "codigo": {
                        "title": "codigo",
                        "description": "codigo del producto",
                        "type": "string"
                    },
                    "nombre": {
                        "title": "nombre",
                        "description": "nombre del producto",
                        "type": "string"
                    },
                    "descripcion": {
                        "title": "descripcion",
                        "description": "descripcion del producto",
                        "type": "string"
                    },
                    "cantidad": {
                        "title": "cantidad",
                        "description": "cantidad del producto",
                        "type": "number"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Producto"
                }
            },
            "ProductoResource": {
                "title": "ProductoResource",
                "description": "Producto resource",
                "properties": {
                    "data": {
                        "title": "data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Producto"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ProductoResource"
                }
            }
        }
    }
}